+++ Engine/Application/cbApplication.h
@@ -130,7 +130,7 @@ namespace eae6320
 			// The default value is a relatively large amount of time;
 			// you may want to experiement with smaller values in your application
 			// and observe the change in responsiveness or simulation accuracy.
-			virtual float GetSimulationUpdatePeriod_inSeconds() const { return 1.0f / 100000.0f; }
+			virtual double GetSimulationUpdatePeriod_inSeconds() const { return 1.0 / 100000.0; }
 
 		private:
 
@@ -184,7 +184,7 @@ namespace eae6320
 			// They will both be called every time the simulation is updated
 			// (i.e. whenever GetSimulationUpdatePeriod_inSeconds() of simulation time has elapsed)
 			virtual void UpdateSimulationBasedOnInput();
-			virtual void UpdateSimulationBasedOnTime(const float i_elapsedSecondCount_sinceLastUpdate);
+			virtual void UpdateSimulationBasedOnTime(const double i_elapsedSecondCount_sinceLastUpdate);
 
 			// Your application should override the following function
 			// to instuct the Graphics system what to render for the next frame
